made all controllers throw errors instead of returning http directly + moved all non-validation errors from controllers to services

+++ src/modules/subevents/services/services.ts
@@ -1,6 +1,7 @@
 import * as Errors from "../../../errors/errors";
 import subeventType from "../../../types/modelTypes/subevent";
 import subeventCreationInputData from "../../../types/subeventInputData";
+import eventParticipationRepo from "../../events/repositories/eventParticipationRepo";
 import eventRepo from "../../events/repositories/eventRepo";
 import subeventParticipationRepo from "../repositories/subeventParticipaitonRepo";
 import subeventRepo from "../repositories/subeventRepo";
@@ -9,6 +10,12 @@ import mongoose from "mongoose";
 const subeventService = {
 
     add: async (input: subeventCreationInputData) => {
+
+        const userIsParticipantInEvent = await eventParticipationRepo.findUserInEvent(input.userId, input.eventId);
+        if (!userIsParticipantInEvent) {
+            throw new Errors.UserIsNotParticipantInEventError(input.eventId);
+        };
+
         const subeventData: subeventType = { ...input }
         const subevent = await subeventRepo.add(subeventData);
         await subeventParticipationRepo.add(input.userId, subevent._id);
@@ -21,11 +28,23 @@ const subeventService = {
     },
 
     getbyEventAndUser: async (userId: mongoose.Types.ObjectId, eventId: mongoose.Types.ObjectId) => {
+
+        const eventExists = await eventRepo.find(eventId);
+        if (!eventExists) {
+            throw new Errors.EventNotFoundError(eventId);
+        };
+
         const subevents = await subeventParticipationRepo.getSubeventByEventAndUser(userId, eventId);
         return { success: true, message: "Subevents retrieved", data: { subevents } };
     },
 
-    signUp: async (userId: mongoose.Types.ObjectId, subeventId: mongoose.Types.ObjectId) => {
+    join: async (userId: mongoose.Types.ObjectId, subeventId: mongoose.Types.ObjectId) => {
+
+        const subeventExists = await subeventRepo.findSubevent(subeventId);
+        if (!subeventExists) {
+            throw new Errors.SubeventNotFoundError(subeventId);
+        };
+
         const alreadySignedUp = await subeventParticipationRepo.findParticipation(userId, subeventId);
         if (alreadySignedUp) {
             throw new Errors.UserAlreadySignedUpToSubeventError(subeventId);
@@ -35,12 +54,24 @@ const subeventService = {
     },
 
     delete: async (subeventId: mongoose.Types.ObjectId) => {
+
+        const subeventExists = await subeventRepo.findSubevent(subeventId);
+        if (!subeventExists) {
+            throw new Errors.SubeventNotFoundError(subeventId);
+        };
+
         const deletedSubevent = await subeventRepo.delete(subeventId);
         await subeventParticipationRepo.deleteAllParticipationsBySubevent(subeventId);
         return { success: true, message: "Subevent deleted", data: { deletedSubevent } }
     },
 
     leave: async (userId: mongoose.Types.ObjectId, subeventId: mongoose.Types.ObjectId) => {
+
+        const subeventExists = await subeventRepo.findSubevent(subeventId);
+        if (!subeventExists) {
+            throw new Errors.SubeventNotFoundError(subeventId);
+        };
+
         await subeventParticipationRepo.deleteOneParticipation(userId, subeventId);
         return { success: true, message: "Participation cancelled" };
     },
